+++ projects/Mallard/src/components/article/index.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { CAPIArticle } from 'src/common'
 import { FlexErrorMessage } from 'src/components/layout/ui/errors/flex-error-message'
 import { color } from 'src/theme/color'
-import { Article } from './types/article'
+import { Article } from './types/article-new'
 import { Crossword } from './types/crossword'
 import { Gallery } from './types/gallery'
 import { Cartoon } from './types/cartoon'
