+++ projects/Mallard/src/components/article/types/article-new.tsx
@@ -1,0 +1,62 @@
+import React, { useEffect, useState } from 'react'
+import { StyleSheet } from 'react-native'
+import { BlockElement } from 'src/common'
+import { OnTopPositionChangeFn } from 'src/screens/article/helpers'
+import { metrics } from 'src/theme/spacing'
+import { Fader } from '../../layout/animators/fader'
+import { ArticleHeaderProps } from '../article-header/types'
+import { PropTypes as StandfirstPropTypes } from '../article-standfirst'
+import { Wrap, WrapLayout } from '../wrap/wrap'
+import { WebviewWithArticle } from './article/webview'
+
+const styles = StyleSheet.create({
+    block: {
+        alignItems: 'flex-start',
+        padding: metrics.horizontal,
+        paddingVertical: metrics.vertical,
+    },
+    webviewWrap: {
+        ...StyleSheet.absoluteFillObject,
+    },
+    webview: {
+        flex: 1,
+        /*
+        The below line fixes crashes on Android
+        https://github.com/react-native-community/react-native-webview/issues/429
+        */
+        opacity: 0.99,
+    },
+})
+
+const Article = ({
+    onTopPositionChange,
+    article,
+    ...headerProps
+}: {
+    article: BlockElement[]
+    onTopPositionChange: OnTopPositionChangeFn
+} & ArticleHeaderProps &
+    StandfirstPropTypes) => {
+    const [wrapLayout, setWrapLayout] = useState<WrapLayout | null>(null)
+    useEffect(() => {
+        onTopPositionChange(false)
+    }, [])
+
+    return (
+        <Fader>
+            {wrapLayout && (
+                <WebviewWithArticle
+                    headerProps={headerProps}
+                    article={article}
+                    scrollEnabled={true}
+                    useWebKit={false}
+                    style={[styles.webview]}
+                    wrapLayout={wrapLayout}
+                />
+            )}
+            <Wrap onWrapLayout={setWrapLayout}></Wrap>
+        </Fader>
+    )
+}
+
+export { Article }
